Update create_test_conda_env.yml
because this `create_test_conda_env.yml` generates the badge(s) and renders the documentation, it should run on pushes to `main`, in addition to `pull_request`. This is needed because, currently AND surprisingly, merging a PR to main doesn't satisfy the `pull_request` criteria. And so the `gh-pages` doc is not always updated when we merge to main